use entrypoints for better compatability on Windows
